requirements.txt
aiogram==2.23.1
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==2.1.1
deta==1.1.0
et-xmlfile==1.1.0
fastapi==0.88.0
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.3
openpyxl==3.0.10
pydantic==1.10.2
python-dotenv==0.21.0
pytz==2022.7
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
tzdata==2022.7
yarl==1.8.2
